requirements.txt
gradio>=4.0.0
openai>=1.0.0
transformers>=4.30.0
torch>=2.0.0
torchaudio>=2.0.0
pypdf2>=3.0.0
python-docx>=0.8.11
reportlab>=4.0.0
datasets>=2.14.0
huggingface-hub>=0.16.0
accelerate>=0.20.0
librosa==0.11.0
